[#30]Refactor(post): type, constant posts에서 사용하도록 정리, API 엔드포인트도 상수로 관리
- post-creation 파일 삭제 후 posts로 통합
- 관리 지점 하나로 통일하기 위해 API 엔드포인트도 상수로 관리
- 불필요한 수정 api 삭제
