+++ types/posts.ts
@@ -1,12 +1,33 @@
 export type PostSortType = "latest" | "reactions" | "comments";
+// 게시글 작성 관련 타입 정의만 포함
 
 export type ConceptType =
-  | "all"
-  | "classical_poetry"
-  | "poetry"
-  | "novel"
-  | "drama"
-  | "essay";
+  | "CLASSICAL_POETRY"
+  | "POETRY"
+  | "NOVEL"
+  | "DRAMA"
+  | "ESSAY";
+
+export type EmotionType =
+  | "HAPPY"
+  | "SAD"
+  | "ANGRY"
+  | "EXCITED"
+  | "CONFUSED"
+  | "PROUD";
+
+// 카테고리 옵션 타입
+export interface CategoryOption {
+  label: string;
+  value: ConceptType;
+}
+
+// 감정 옵션 타입
+export interface EmotionOption {
+  label: string;
+  value: EmotionType;
+  color: string;
+}
 
 interface ReactionDetailCountMap {
   happy: number;
@@ -60,3 +81,17 @@ export interface PostQueryParams {
   limit?: number;
   keyword?: string;
 }
+
+export type CreatePostRequest = {
+  title: string;
+  content: string;
+  conceptType: string;
+  emotionType: string | null;
+  imageUrl?: string;
+};
+
+export type CreatePostResponse = {
+  postId: number;
+  afterTitle: string;
+  afterContent: string;
+};
